+++ src/minigraphs/__init__.py
@@ -8,7 +8,7 @@ Python package for experimentation with network models
 _submodules = [
     "simulation",
     "data",
-    "miniaturize"
+    "mcmc"
     "sampling"
 ]
