+++ src/minigraphs/mcmc/__init__.py
@@ -1,0 +1,5 @@
+from .annealer import SimulatedAnnealing
+
+__all__ = [
+    "SimulatedAnnealing"
+]
