+++ src/minigraphs/simulation/sir.py
